e2e: acceptance テストを一箇所に統合
US-01 シナリオ 1 を関数直呼びから page.goto() + request.post('/upload-image') による
実 HTTP テストに書き換え。e2e/editor-image-upload.spec.ts は重複のため削除。

- シナリオ 1: HTTP テスト（エンドポイント経由で変換を検証）
- シナリオ 2・3, US-02: 関数テストのまま維持（シナリオ 3 はサーバー設定の都合で HTTP 不可）
